requirements.txt
Django==1.7.1
Pygments==2.0.1
dj-database-url==0.3.0
dj-static==0.0.6
django-filter==0.9.1
django-toolbelt==0.0.1
djangorestframework==3.0.0
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
wsgiref==0.1.2
